+++ Frontend/src/components/ErrorModal.js
