+++ Frontend/src/pages/Login.css
@@ -12,6 +12,7 @@
   width: 400px;
   height:400px;
   padding: 30px;
+  align-items: center;
   border-radius: 10px;
   background-color: #fff;
   box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
@@ -35,30 +36,29 @@
 
 /* 입력 필드 */
 .login-input {
-  width: 93%;
+  width: 90%;
   padding: 12px;
-  margin: 8px 0;
+  margin: 12px 0;  /* ✅ 기존 8px → 12px로 증가 (간격 추가) */
   border: 1px solid #ddd;
   border-radius: 5px;
-  font-size: 16px;
-  margin-top:25px;
 }
 
 /* 메인 로그인 버튼 (네이버 스타일) */
 .login-main-button {
-  width: 100%;
+  width: 97%;
   padding: 12px;
-  background-color: #f27f4e; /* 네이버 초록색 */
+  background-color: #5A81FA; /* 네이버 초록색 */
   color: white;
   border: none;
   border-radius: 5px;
   font-size: 16px;
   cursor: pointer;
-  margin-top: 50px;
+  margin-top: 80px;
+  margin-bottom:25px;
 }
 
 .login-main-button:hover {
-  background-color: #e48238;
+  background-color: #4c64f0;
 }
 
 .login-button:hover {
