--- src/components/Footer.css
@@ -1,29 +1,0 @@
-/* 푸터 스타일 */
-.footer {
-    width: 100%;
-    background-color: #f8f9fa;
-    padding: 15px 0;
-    text-align: center;
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    font-size: 14px;
-    color: #6c757d;
-  }
-  
-  /* 푸터 링크 스타일 */
-  .footer-links {
-    margin-top: 5px;
-  }
-  
-  .footer-links a {
-    text-decoration: none;
-    color: #0077b6;
-    margin: 0 10px;
-    font-weight: 500;
-  }
-  
-  .footer-links a:hover {
-    text-decoration: underline;
-  }
-
