--- src/components/Footer.js
@@ -1,17 +1,0 @@
-import React from "react";
-import "./Footer.css";
-
-function Footer() {
-  return (
-    <footer className="footer">
-      <p>© 2025 여행지 소개글 LLC All rights reserved.</p> 
-      <div className="footer-links">
-        <a href="#">이용 약관</a>
-        <a href="#">개인정보 정책</a>
-        <a href="#">사이트맵</a>
-      </div>
-    </footer>
-  );
-}
-
-export default Footer;
