--- src/components/Header.css
@@ -1,84 +1,0 @@
-/* 헤더 스타일 */
-.header {
-  width: 100%;
-  background-color: white;
-  box-shadow: 0px 2px 5px rgba(0, 0, 0, 0.1);
-  padding: 10px 0; /* 헤더 높이를 조정하여 로고가 중앙에 오도록 함 */
-  position: fixed;
-  top: 0;
-  left: 0;
-  z-index: 1000;
-}
-
-/* 헤더 내부 컨테이너 */
-.header-container {
-  max-width: 1200px;
-  margin: 0 auto;
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  height: 60px; /* 헤더 높이를 조정하여 일관된 디자인 유지 */
-}
-
-/* 로고 컨테이너 */
-.logo {
-  display: flex;
-  align-items: center; /* 로고와 텍스트 수직 중앙 정렬 */
-  gap: 10px; /* 로고와 텍스트 사이 여백 */
-}
-
-/* 로고 이미지 */
-.logo img {
-  height: 80px; /* 로고 크기 줄이기 */
-  width: auto;
-  border-radius: 10%;
-}
-
-/* 로고 텍스트 */
-.logo-text {
-  font-size: 22px; /* 글자 크기 */
-  font-weight: bold;
-  color: #222; /* TripAdvisor처럼 어두운 색 */
-  font-family: "Arial", sans-serif; /* 깔끔한 글꼴 */
-}
-
-/* 검색창 */
-.search-bar {
-  flex: 1;
-  display: flex;
-  justify-content: center;
-}
-
-.search-bar input {
-  width: 70%;
-  padding: 10px 15px;
-  border: 1px solid #ddd;
-  border-radius: 25px;
-  outline: none;
-}
-
-/* 네비게이션 메뉴 */
-.nav-menu {
-  display: flex;
-  align-items: center;
-}
-
-.nav-menu a {
-  text-decoration: none;
-  color: black;
-  margin-left: 20px;
-  font-size: 14px;
-  font-weight: 500;
-}
-
-.nav-menu a:hover {
-  color: #0077b6;
-}
-
-/* 언어 선택 */
-.language {
-  margin-left: 20px;
-  font-size: 14px;
-  font-weight: bold;
-  cursor: pointer;
-}
