--- src/pages/Login.css
@@ -1,88 +1,0 @@
-/* 로그인 페이지 전체 컨테이너 */
-.login-container {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  height: 90vh;
-}
-
-/* 로그인 컨테이너 박스 */
-.login-box {
-  width: 400px;
-  height:400px;
-  padding: 30px;
-  align-items: center;
-  border-radius: 10px;
-  background-color: #fff;
-  box-shadow: 0 2px 10px rgba(0, 0, 0, 0.1);
-  text-align: center;
-  margin-top:70px;
-}
-
-/* 로고 스타일 */
-.login-logo {
-  height: 60px; /* 로고 크기 조정 */
-  width: auto;
-  margin-bottom: 20px;
-}
-
-/* 로그인 제목 */
-.login-title {
-  font-size: 25px;
-  font-weight: bold;
-  margin-bottom: 20px;
-}
-
-/* 입력 필드 */
-.login-input {
-  width: 90%;
-  padding: 12px;
-  margin: 12px 0;  /* ✅ 기존 8px → 12px로 증가 (간격 추가) */
-  border: 1px solid #ddd;
-  border-radius: 5px;
-}
-
-/* 메인 로그인 버튼 (네이버 스타일) */
-.login-main-button {
-  width: 97%;
-  padding: 12px;
-  background-color: #5A81FA; /* 네이버 초록색 */
-  color: white;
-  border: none;
-  border-radius: 5px;
-  font-size: 16px;
-  cursor: pointer;
-  margin-top: 80px;
-  margin-bottom:25px;
-}
-
-.login-main-button:hover {
-  background-color: #4c64f0;
-}
-
-.login-button:hover {
-  background-color: #e9ecef;
-}
-
-/* 아이디 찾기, 비밀번호 찾기, 회원가입 */
-.login-links {
-  margin-top: 15px;
-  font-size: 14px;
-}
-
-.login-links a {
-  color: #0077b6;
-  text-decoration: none;
-}
-
-.login-links a:hover {
-  text-decoration: underline;
-}
-
-/* 이용 약관 */
-.login-footer {
-  margin-top: 20px;
-  font-size: 12px;
-  color: gray;
-}
