--- src/tailwind.config.js
@@ -1,8 +1,0 @@
-module.exports = {
-    content: ["./src/**/*.{js,jsx,ts,tsx}"],  // ✅ Tailwind 적용될 파일 경로 설정
-    theme: {
-      extend: {},
-    },
-    plugins: [],
-  };
-
